Migrates CurrencyUtils formatSigFig to Settings
