--- src/main/java/com/maldahleh/stockmarket/utils/CurrencyUtils.java
@@ -1,30 +1,0 @@
-package com.maldahleh.stockmarket.utils;
-
-import java.math.BigDecimal;
-import lombok.experimental.UtilityClass;
-
-@UtilityClass
-public class CurrencyUtils {
-
-  public String formatSigFig(Long input, String unknown) {
-    if (input == null) {
-      return unknown;
-    }
-
-    return sigFigNumber(input);
-  }
-
-  public String sigFigNumber(BigDecimal input) {
-    return sigFigNumber(input.doubleValue());
-  }
-
-  public String sigFigNumber(double input) {
-    String suffixes = "kmbt";
-    if (input < 1000) {
-      return String.valueOf(input).replace(".0", "");
-    }
-
-    int exponent = (int) (Math.log(input) / Math.log(1000));
-    return String.format("%.1f%c", input / Math.pow(1000, exponent), suffixes.charAt(exponent - 1));
-  }
-}
